+++ CHANGELOG.md
@@ -48,6 +48,9 @@
 # 0.51.0 [unreleased]
 
 - Count bandwidth at the application level. Previously `BandwidthLogging` would implement `Transport` and now implements `StreamMuxer` ([PR 3180](https://github.com/libp2p/rust-libp2p/pull/3180)).
+    - `BandwidthLogging::new` now requires a 2nd argument: `Arc<BandwidthSinks>`
+    - Remove `BandwidthFuture`
+    - Rename `BandwidthConnecLogging` to `InstrumentedStream`
 
 # 0.50.0